+++ backend/src/main/java/com/alaorden/service/impl/AddressServiceImpl.java
@@ -98,7 +98,7 @@ public class AddressServiceImpl implements AddressService {
         List<Location> selectedLocations = new ArrayList<>();
 
         for(Location location : locations){
-            double dist = computeDistance(latitud,longitud,location.getLatitude().doubleValue(),location.getLongitude().doubleValue());
+            double dist = computeDistance(latitud,longitud,location.getLatitude(),location.getLongitude());
             if(dist <= 500.5)//<---- Distancia minima de 500.5 kilometros
             {
                 if(!Exist(location.getFranchise().getName(),setLocations)){
